change global temperatures from float to int16_t, that way we can do the casting (and rounding) in only one location

+++ src/display.h
@@ -3,5 +3,5 @@
 #include <Arduino.h>
 
 void displayInit();
-void setDisplay(float value);
+void setDisplay(int16_t value);
 void updateUnits();
